Removed previous stuff about solution annotations and added a more general concept of adding annotations to both nodes and solutions.
Annotations are now just objects of any type that can be associated with
a node via the event bus system. The ORGraphSearch subscribes to that
kind of events and only needs to be registered by classes that want to
emit this kind of information.

Perhaps it would be better to replace the event bus system by some other
technique, because it is very specific for this search, and perhaps we
have an over-generalization here.