The recorder got the functionality of storing tooltips.
To achieve this, a list is stored, in which every tooltip from a node is stored. For loading this list
is recreated and gets bound to the corresponding node. The nodes are recognised by a unique id, which lead
to the usage of AbstractNode. Therefore it is recommeneded to use the AbstractGraphGenerator as a wrapper for
the main graphGenerator.