Intermediate Result Cache
ReproducibleInstances are new Instances which can be used to track the creation of Instances by saving the origin and splits.
Added intermediate result cache. Computed loss values can be stored in a database now. See CacheEvaluatorMeasureBridge.
The Cache can be used to store results and later reproduce them or to also use cached results in the search.
If a normal Instances Object is used the cache is not activated. Flags to control the use of the cache can be set in the ReproducibleInstances.