Added NumericDataset which stores each attribute value as a double.
To this end, simple attribute types were introduced, which can be used
to still identify which attributes are nominal or ordinal etc.

Added interface for learning algorithms which is generic in the input
and output space as well as the type of config with which the respective
algorithm can be configured. Maybe the name is still not as appropriate.